requirements.txt
python-dotenv
openai
easyocr
fuzzywuzzy
matplotlib
python-Levenshtein
reportlab